mdserver_remote: use short timeout to check server for latest TLF
This waits 500ms for a server response before giving up and using the
cached MD if there is one.

In most cases `folderBranchOps` will roll forward to the latest
revision without this fix, but in the meantime bad stuff can happen:
if it's an unsynced TLF, the user can get "block not found" errors, or
rekeys/upgrades can fail because of conflicts.

Issue: KBFS-3665
Issue: KBFS-3676